mm,page_alloc,cma: introduce a customisable threshold for allocating pages in cma
commit  fc5033ff7f7278510c5060b99ce9a6d46bfcc24b from
https://github.com/raspberrypi/linux.git rpi-6.1.y

On some platforms the cma area can be half the entire system memory,
meaning that allocations start happening in the cma area immediately.
This leads to fragmentation and subsequent fatal cma_alloc failures.

We introduce an "alloc_in_cma_threshold" parameter which requires that
this many sixteenths of the free pages must be in cma before it will
try to use them. By default this is set to 12, but the previous
behaviour can be restored by setting it to 8 on startup.

Signed-off-by: David Plowman <david.plowman@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>