drm/panel: panel-ilitek9881c: Add prepare_upstream_first flag
commit  267c4dc7476704478365ee109fe143f31cde81a6 from
https://github.com/raspberrypi/linux.git rpi-5.18.y

The panel sends MIPI DCS commands during prepare and is expecting
the bus to remain in LP-11 state in-between.

Set the prepare_upstream_first flag so that the upstream DSI host
is prepared / pre_enabled first, and therefore the bus is in a
defined state.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>