drm/panel: panel-ilitek9881c: Add prepare_upstream_first flag
commit  423384b965af8dd8dc75270de43757f9013f9681 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

The panel sends MIPI DCS commands during prepare and is expecting
the bus to remain in LP-11 state in-between.

Set the prepare_upstream_first flag so that the upstream DSI host
is prepared / pre_enabled first, and therefore the bus is in a
defined state.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>