drm/vc4: Force trigger of dlist update on margins change
commit  454cc93a970483c57f91934ef03b0d4c1ef5de57 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

When the margins are changed, the dlist needs to be regenerated
with the changed updated dest regions for each of the planes.

Setting the zpos_changed flag is sufficient to trigger that
without doing a full modeset, therefore set it should the
margins be changed.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>