drm/vc4: Force trigger of dlist update on margins change
commit  98d05c6d95f0c08a1b0cb8415ddba4960ff87526 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

When the margins are changed, the dlist needs to be regenerated
with the changed updated dest regions for each of the planes.

Setting the zpos_changed flag is sufficient to trigger that
without doing a full modeset, therefore set it should the
margins be changed.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>