drm/atomic: If margins are updated, update all planes.
commit  aae9c2b8838c0278590afef52664346a5716ba23 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

Margins may be implemented by scaling the planes, but as there
is no way of intercepting the set_property for a standard property,
and all planes are checked in drm_atomic_check_only before the
connectors, there's now way to add the planes into the state
from the driver.

If the margin properties change, add all corresponding planes to
the state.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>