ARM: dts: bcm2835: Switch HSM clock to firmware
commit  d9ab5e133382fcd765f5ad9dc6e8274a81c2f8fc from
https://github.com/raspberrypi/linux.git rpi-5.18.y

When the bootloader boots without a display connected to the
HDMI connector, it won't initialize the HSM clock.

Then, when Linux boots, the default parent fetched by the clock
framework will be the first parent for the HSM clock, which is "gnd".

That clock is never registered, leading to the HSM clock being orphaned,
which leads to other issues with the recent clock range handling.

It's not clear what the proper solution would be, but the firmware
clocks driver doesn't have that parenting issue and the clock rate is
properly initialized for this clock whenever it hasn't been set by the
bootloader.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>