vc04_services: bcm2835-codec: Set port format from s_selection
commit  6d1a0b05b56479b25c956769e91cd70d71a94a0d from
https://github.com/raspberrypi/linux.git rpi-5.19.y

s_selection allows the crop region of an uncompressed pixel
format to be specified, but it wasn't passing the setting on to
the firmware. Depending on call order this would potentially
mean that the crop wasn't actioned.

Set the port format on s_selection if we have a component created.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>