drm: vc4: Add 3:3:2 and 4:4:4:4 RGB/RGBX/RGBA formats
commit  dcbeba34547e7fac7359216ee063465c4411a5fb from
https://github.com/raspberrypi/linux.git rpi-6.1.y

The hardware supports the 332 8bpp and 4:4:4:4 16bpp formats,
but the table of supported formats didn't include them.
Add them in.

In theory they are supported for T-format as well as linear,
but without a way to test them just add them as linear for now.

Suggested-by: vrazzer <teamvraz@pipmail.net>
Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>