ARM: dts: Adopt the upstream snd_bcm2835 handling
commit  0b6de3a112e97a76317d6d6f81da137bb5355893 from
https://github.com/raspberrypi/linux.git rpi-5.19.y

Swimming against the upstream current when it comes to the snd_bcm2835
audio driver has become increasingly difficult. This is an attempt to
convert our current audio handling to the new scheme.

The snd_bcm2835 driver should no longer be instantiated by Device Tree,
which means it loses the ability to use DT properties for its
configuration. Instead, enabling and disabling audio outputs - HDMI and
headphones (if present) - should be done by module parameters, which
can be supplied on the kernel command line.

There may be the potential for problems if the "audio" dtparam is used
_after_ an audio- related overlay is applied, but otherwise it should
be OK.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>