dtoverlays: Use PDAF variant for imx258 overlay
commit  9cca268fc3fe63394eec1ab03e11f7faf4380c9f from
https://github.com/raspberrypi/linux.git rpi-5.19.y

There are a number of variants of the imx258 sensor. The only
version that is currently made into a module for use with the Pi
is from Vision Components, and that supports PDAF and needs the
appropriate register configuration.

Update the compatible string accordingly.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>