dtoverlays: Reshuffle image sensor overlays to allow use with muxes
commit  d68f1175b1a59a98392cfa4171d746db26510d9c from
https://github.com/raspberrypi/linux.git rpi-6.0.y

There are third-party CSI-2 muxes available that ideally want to
be configured to have any of the supported image sensors.
That wants the sensor configuration to be split out into fragements
that can be included in the relevant places, and then an overall
overlay that pulls them in.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>