pnmtologo: use relocatable file name
The logo generation utility is capturing the source of the logo
in the generated .c file. The source file is absolute (as passed
by make), so the full path is captured.

This makes the source fail reproducibility tests.

We use basename() to just get the source file name, and use
that in the generated .c file.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>