media: i2c: imx290: Updating VBLANK should update exposure in all states
commit  6667f55124227fb9029f0d9e79516c368ff9a922 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

The code to update the range of the exposure control was after a
check that the sensor was powered up. This is incorrect as the
range should be updated under all conditions.

Move the range update code to the correct place.

Fixes: 52d076ea221e "media: i2c: imx290: Handle exposure correctly when vblank changes"

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>