firmware: raspberrypi: Provide a helper to query a clock max rate
commit  2d1bb71f92638ba6948b7c67b3672ef6c9290841 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

The firmware allows to query for its clocks the operating range of a
given clock. We'll need this for some drivers (KMS, in particular) to
infer the state of some configuration options, so let's create a
function to do so.

Acked-by: Stephen Boyd <sboyd@kernel.org>
Reviewed-by: Florian Fainelli <f.fainelli@gmail.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>