vc04_services: bcm2835-codec: Allow encoder_cmd on ISP and deinterlace
commit  80aaa961c308a680f34157f1397c13b7cee45e04 from
https://github.com/raspberrypi/linux.git rpi-6.0.y

ISP and deinterlace also need a mechanism for passing effectively
an EOS through the pipeline to signal when all buffers have been
processed.

VIDIOC_ENCODER_CMD does exactly this for encoders, so reuse the same
function for ISP and deinterlace.
(VIDIOC_DECODER_CMD is slightly different in that it also passes
details of when and how to stop, so is not as relevant).

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>