firmware: raspberrypi: Introduce rpi_firmware_find_node()
commit  37c0d9f797f3008e9e7adfa68c6f436c46645049 from
https://github.com/raspberrypi/linux.git rpi-6.1.y

A significant number of RaspberryPi drivers using the firmware don't
have a phandle to it, so end up scanning the device tree to find a node
with the firmware compatible.

That code is duplicated everywhere, so let's introduce a helper instead.

Link: https://lore.kernel.org/r/20220815-rpi-fix-4k-60-v4-1-a1b40526df3e@cerno.tech
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>