build: partially revert kbuild: add phony targets to ./Kbuild
As part of commit ed7ceac157c27bdc64e79a322 [kbuild: add phony targets to ./Kbuild]
a loop has been created between prepare0 and prepare.

prepare depends on prepare0, and prepare0 now explictity calls "make
prepare", and our loop continues until an eventual hang.

We keep the rest of the commit, but drop the explicit call to "make
prepare", and we restore the previous functionality which we use to
recreate the kernel build support applications on the target.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>