build: make 'prepare' conditional on command line
If something like "make scripts prepare" is used on the command
line, we end up with an infinite loop of dependencies as the
prepare0: target explicitly calls 'make prepare'.

To avoid this, we check if prepare has already been passed as
an explicit target/goal and drop it from the explicit make
call.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>