drm/vc4: crtc: Make encoder lookup helper public
commit  1a817234061ba78cb70f8e94c9e7a933ed70ea8c from
https://github.com/raspberrypi/linux.git rpi-6.2.y

We'll need a function that looks up an encoder by its vc4_encoder_type.
Such a function is already present in the CRTC code, so let's make it
public so that we can reuse it in the unit tests.

Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Maíra Canal <mcanal@igalia.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>