drm/tests: helpers: Make sure the device is bound
commit  b78a2eae4b7cc10c266e14c959cdca3d7f1b8360 from
https://github.com/raspberrypi/linux.git rpi-6.2.y

The device managed resources are freed when the device is detached, so
it has to be bound in the first place.

Let's create a fake driver that we will bind to our fake device to
benefit from the device managed cleanups in our tests.

Reviewed-by: Maíra Canal <mcanal@igalia.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Signed-off-by: Meng Li <Meng.Li@windriver.com>