drm/vc4: Add TXP encoder type
commit  89485a32f3e607d60a1ef795bb4879e5b2415a00 from
https://github.com/raspberrypi/linux.git rpi-6.2.y

The TXP is integrated as a separate CRTC/Encoder/Connector combo, but
for some reason doesn't rely on the vc4_encoder type and it's associated
type.

Let's create a type to make it consistent with the other encoders.

Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Signed-off-by: Maxime Ripard <maxime@cerno.tech>
Link: https://patchwork.freedesktop.org/patch/msgid/20221123-rpi-kunit-tests-v1-15-051a0bb60a16@cerno.tech
Signed-off-by: Meng Li <Meng.Li@windriver.com>