media: dw9807-vcm: Add support for DW9817 bidirectional VCM driver
commit  8d4906085896047ad59f30892279cbc885085fe4 from
https://github.com/raspberrypi/linux.git rpi-6.2.y

The DW9817 is effectively the same as DW9807 from a programming
interface, however it drives +/-100mA instead of 0-100mA. This means
that the power on ramp needs to take the lens from the midpoint, and
power off return it there. It also changes the default position for
the module.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>