media: i2c: imx290: Access link_freq_index directly
commit  296568cbc68884118b16315cc3f54f6bdc19c2cb from
https://github.com/raspberrypi/linux.git rpi-6.2.y

The imx290_get_link_freq_index() function hides the fact that it relies
on the imx290 current_mode field, which obfuscates the code instead of
making it more readable. Inline it in the callers, and use the mode
pointer we already have in imx290_ctrl_update() instead of using the
current_mode field.

Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Reviewed-by: Alexander Stein <alexander.stein@ew.tq-group.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>